- resolve #215: Update Xamarin sample's dependencies to Rollbar.NET v2
