+++ Sample.Xamarin.Forms/Sample.Xamarin.Forms.Android/Properties/AndroidManifest.xml
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android" android:versionCode="1" android:versionName="1.0" package="com.companyname.Sample.Xamarin.Forms" android:installLocation="auto">
-	<uses-sdk android:minSdkVersion="24" android:targetSdkVersion="27" />
+	<uses-sdk android:minSdkVersion="24" android:targetSdkVersion="28" />
 	<uses-permission android:name="android.permission.INTERNET" />
 	<uses-permission android:name="com.android.launcher.permission.INSTALL_SHORTCUT" />
 	<application android:label="Sample.Xamarin.Forms.Android" android:icon="@drawable/icon"></application>
