+++ Sample.Xamarin.Forms/Sample.Xamarin.Forms.Android/Sample.Xamarin.Forms.Android.csproj
@@ -55,7 +55,7 @@
     <PackageReference Include="System.Memory">
       <Version>4.5.1</Version>
     </PackageReference>
-    <PackageReference Include="Xamarin.Forms" Version="3.3.0.912540" />
+    <PackageReference Include="Xamarin.Forms" Version="3.4.0.1008975" />
     <PackageReference Include="Xamarin.Android.Support.Design" Version="27.0.2.1" />
     <PackageReference Include="Xamarin.Android.Support.v7.AppCompat" Version="27.0.2.1" />
     <PackageReference Include="Xamarin.Android.Support.v4" Version="27.0.2.1" />
