+++ Sample.Xamarin.Forms/Sample.Xamarin.Forms.iOS/AppDelegate.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-
+using System.Threading.Tasks;
 using Foundation;
+using Rollbar;
 using UIKit;
 
 namespace Sample.Xamarin.Forms.iOS
