+++ SolutionCommon.csproj
@@ -14,6 +14,7 @@
       Improvements and fixes:
       - resolve #213: Improve Rollbar.Benchmarker configurability and coverage
       - resolve #214: Update BenchmarkDotNet dependency to the latest
+      - resolve #215: Update Xamarin sample's dependencies to Rollbar.NET v2
       
     </PackageReleaseNotes>
